♻️ [Refactor]: feature/study/schedule 생성 & Widget/study/schedule 구조에 맞게 수정
