+++ src/app/(auth)/study-room/[id]/layout.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import ImportantNoticeBar from "@/entities/study/notice/ui/ImportantNoticeBar";
 import { getStudy } from "@/features/studies/services/studies.service";
-import StudyRoomIntendedCard from "@/features/study-room/component/StudyRoomIntendedCard";
+import ScheduleIntendedCard from "@/widget/study-room/schedule/ui/ScheduleIntendedCard";
 import StudyRoomSideBarCard from "@/features/study-room/component/StudyRoomSideBarCard";
 import StudyRoomTitleCard from "@/features/study-room/component/StudyRoomTitleCard";
 import Grid from "@/shared/components/atoms/Grid";
@@ -21,7 +21,7 @@ const layout = async ({ children, params }: StudyRoomLayoutProps) => {
       <ImportantNoticeBar />
       <Grid cols={12} gap={5}>
         <StudyRoomTitleCard data={studyDetailData} />
-        <StudyRoomIntendedCard />
+        <ScheduleIntendedCard />
         <StudyRoomSideBarCard />
         {children}
       </Grid>
